Add a message box in case of no search results on homepage
This commit slightly enhances search experience, by displaying a message
"No results found..." for case whenever a query in a search box is not
found on the available list of entries. This will be useful as the user
will not be waiting for search results in case the csv source datasheet does not
include it.